Port supervisord to pytest
