--- supervisord/test/ci/resources/program_0.sh
@@ -1,7 +1,0 @@
-# (C) Datadog, Inc. 2010-2017
-# All rights reserved
-# Licensed under Simplified BSD License (see LICENSE)
-
-# dummy program that runs for 10 seconds and dies
-echo 'test' >> $VOLATILE_DIR/supervisor/started_0
-sleep 10
