--- supervisord/test/ci/resources/supervisord.conf
@@ -1,31 +1,0 @@
-[unix_http_server]
-file=VOLATILE_DIR/supervisor.sock
-
-[rpcinterface:supervisor]
-supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
-
-[supervisorctl]
-serverurl=unix://VOLATILE_DIR/supervisor/supervisor.sock
-
-[supervisord]
-logfile=VOLATILE_DIR/supervisor/supervisord.log
-logfile_maxbytes=50MB
-logfile_backups=10
-loglevel=info
-pidfile=VOLATILE_DIR/supervisor/supervisord.pid
-childlogdir=VOLATILE_DIR/supervisor
-
-[program:program_0]
-command=/bin/sh VOLATILE_DIR/supervisor/program_0.sh
-autostart=true
-autorestart=false
-
-[program:program_1]
-command=/bin/sh VOLATILE_DIR/supervisor/program_1.sh
-autostart=true
-autorestart=false
-
-[program:program_2]
-command=/bin/sh VOLATILE_DIR/supervisor/program_2.sh
-autostart=true
-autorestart=false
