+++ supervisord/tests/compose/supervisord.yaml
@@ -1,0 +1,7 @@
+version: '3.5'
+
+services:
+  supervisord:
+    image: "datadog/docker-library:supervisord_3_3_3"
+    ports:
+      - "19001:19001"
